refactor(routes): move update_routine HTTP + MCP endpoints into routes/update_routine
Follows the routes/health / routes/shutdown / routes/restart /
routes/get_lock_status / routes/list_agents / routes/cleanup_workbenches /
routes/list_routines / routes/get_routine / routes/delete_routine /
routes/create_routine / routes/list_routine_runs template: splits the
PATCH /routines/{id} handler (and its PUT alias, previously routines::update
and routines::replace in src/routines/handlers.rs) and the MCP update_routine
tool into their own folder sharing a single logic::build().
