+++ src/openapi.rs
@@ -26,8 +26,8 @@
         crate::routes::create_routine::create_routine,
         crate::routes::get_routine::get_routine,
         crate::routines::get_prompt_preview,
-        crate::routines::replace,
-        crate::routines::update,
+        crate::routes::update_routine::replace,
+        crate::routes::update_routine::update_routine,
         crate::routes::delete_routine::delete_routine,
         crate::routines::trigger,
         crate::routines::scheduled_trigger,
