+++ src/routes/update_routine/mcp_tests.rs
@@ -1,0 +1,40 @@
+#![allow(
+    clippy::missing_docs_in_private_items,
+    reason = "test helpers and fixtures do not need doc comments"
+)]
+
+use super::MoadimMcp;
+use crate::routes::mcp::mcp_types::UpdateRoutineInput;
+
+fn make_handler() -> MoadimMcp {
+    MoadimMcp::new(
+        crate::routines::new_store(),
+        crate::paths::routines_dir(),
+        0,
+        std::sync::Arc::new(tokio::sync::Notify::new()),
+    )
+}
+
+#[test]
+fn update_routine_tool_not_found_is_error() {
+    use rmcp::handler::server::wrapper::Parameters;
+    let handler = make_handler();
+    let result = handler
+        .update_routine(Parameters(UpdateRoutineInput {
+            id: "no-such".into(),
+            schedule: None,
+            title: Some("x".into()),
+            agent: None,
+            model: None,
+            prompt: None,
+            goal: None,
+            repositories: None,
+            machines: None,
+            enabled: None,
+            ttl_secs: None,
+            max_runtime_secs: None,
+            tags: None,
+        }))
+        .unwrap();
+    assert!(result.is_error.unwrap_or(false));
+}
