+++ src/routes/update_routine/mod.rs
@@ -1,0 +1,12 @@
+//! Update routine: shared logic (`logic.rs`), HTTP handlers (`http.rs`), and MCP tool (`mcp.rs`,
+//! declared from `routes::mcp` so it can reach `MoadimMcp`'s private state).
+
+pub(crate) mod logic;
+
+#[path = "http.rs"]
+mod http;
+#[allow(
+    unused_imports,
+    reason = "utoipa's OpenApi derive resolves these hidden __path_* types via crate::routes::update_routine::__path_*, generated by #[utoipa::path] on the re-exported handlers below"
+)]
+pub use http::{__path_replace, __path_update_routine, replace, update_routine};
